refactor: moved jstart and mbox to jamJSTranslator
