+++ lib/ohm/jamscript/jam.js
@@ -46,14 +46,8 @@ function compile(cInput, jsInput, lineNumber, yieldPoint) {
 
   return {
     C: cResults.C,
-    JS:
-      jsResults.JS.requires +
-      "\nfunction userProgram() {" +
-      preamble +
-      cResults.JS +
-      jsResults.JS.jsout +
-      '\n}\njworklib.run(function() { console.log("JAMLib 1.0beta Initialized."); userProgram(); } );\n',
-    jstart: cResults.jstart,
+    JS: jsResults.JS.jsout,
+    jstart: jsResults.JS.jstart,
     annotated_JS: jsResults.JS.jsout + cResults.JS,
     maxLevel: jsResults.maxLevel,
     hasJdata: jsResults.hasJdata,
