+++ lib/ohm/jamscript/jamCTranslator.js
@@ -779,40 +779,11 @@ function translate(tree) {
   results.JS += jsout.JS;
   results.annotated_JS += jsout.annotated_JS;
 
-  var mbox = activities.generateMbox(symbolTable.activities.js);
-  results.JS = mbox + results.JS;
-  results.jstart = generateJStart(mbox);
-  results.annotated_JS += mbox;
+
 
   return results;
 }
 
-function generateJStart(mBoxDefinitions = "", conditionDefinitions = "") {
-  return `
-  const minicore = require('./core/minicore');
-  const jaminit = require('./core/jaminit');
-  const JAMCore = require('./core/jamcore');
-
-  ${mBoxDefinitions}
-  ${conditionDefinitions}
-
-  async function launch()
-  {
-      minicore.init('./jamout.js', './dummysched.js', []);
-      var ports = await minicore.run();
-      var jsys = await jaminit.run();
-      var jcore = new JAMCore(jsys);
-      jcore.registerFuncs(mbox);
-      jcore.registerConds(conds);
-      await jcore.run();
-      jcore.addWorker(ports.app);
-  }
-
-  launch().then(()=> {
-      console.log("Starting App..");
-  });
-  `;
-}
 
 function offsetNumber(numMatch, offset) {
   return parseInt(numMatch) - offset;
