+++ lib/ohm/jamscript/jamJSTranslator.js
@@ -66,22 +66,33 @@ var jamJSTranslator = {
       }
     }
 
-    var requires = "";
-    requires += "const Worker = require('tiny-worker');\n";
-    requires += "const jworklib = require('jamserver/jworklib');\n";
-    if (hasJdata) {
-      requires += "const JAMManager = require('jamserver/jammanager');\n";
-      requires += "const JAMLogger = require('jamserver/jamlogger');\n";
-      requires +=
-        "const JAMBroadcaster = require('jamserver/jambroadcaster');\n";
-    }
-    requires += "var jsys;\n";
-    requires += "var jman;\n";
+    const requires = `
+    const {
+        workerData, BroadcastChannel
+    } = require('worker_threads');
+    const globals = require('./utils/constants').globals;
+    const worklib = require('./modules/jworklib');
+    `;
 
-    annotated_JS = requires + annotated_JS;
+    const programInitiator = `
+    (function initApplication(parent)
+    {
+        const appchannel = new BroadcastChannel(globals.ChannelName.APP_LIBRARY);    
+        parent.postMessage({cmd: "READY"});
+        worklib.init(workerData.port, appchannel);
+        worklib.run();
+    })(workerData.port);  
+    `;
+
+    var mbox = activities.generateMbox(symbolTable.activities.js);
+    
+    annotated_JS = requires + programInitiator + annotated_JS;
 
     return {
-      JS: { requires: requires, jsout: jsout },
+      JS: {
+        jsout: requires + programInitiator + mbox + jsout,
+        jstart: generateJStart(mbox)
+      },
       annotated_JS: annotated_JS,
       maxLevel: maxLevel,
       hasJdata: hasJdata,
@@ -700,6 +711,33 @@ function getFunctionNames(node) {
 
   var matches = code.match(functionCallExtractionPattern);
   return matches.map((match) => match.substring(0, match.indexOf("(")));
+}
+
+function generateJStart(mBoxDefinitions = "", conditionDefinitions = "") {
+  return `
+  const minicore = require('./core/minicore');
+  const jaminit = require('./core/jaminit');
+  const JAMCore = require('./core/jamcore');
+
+  ${mBoxDefinitions}
+  ${conditionDefinitions}
+
+  async function launch()
+  {
+      minicore.init('./jamout.js', './dummysched.js', []);
+      var ports = await minicore.run();
+      var jsys = await jaminit.run();
+      var jcore = new JAMCore(jsys);
+      jcore.registerFuncs(mbox);
+      jcore.registerConds(conds);
+      await jcore.run();
+      jcore.addWorker(ports.app);
+  }
+
+  launch().then(()=> {
+      console.log("Starting App..");
+  });
+  `;
 }
 
 es5Translator.FunctionDeclaration = function (
